fix($LottoGenerator) : 로또 생성기 인터페이스 화
lottoGenerator -> 인터페이스 화
NumberGenerator -> 구현
구조로 변경
