+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,14 +1,9 @@
 package lotto.domain;
 
-import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 
-public class LottoGenerator {
-    private static final int RANDOM_NUMBER_START = 1;
-    private static final int RANDOM_NUMBER_END = 45;
-    private static final int RANDOM_NUMBER_COUNT = 6;
-
-    public List<Integer> generate(){
-        return Randoms.pickUniqueNumbersInRange(RANDOM_NUMBER_START,RANDOM_NUMBER_END,RANDOM_NUMBER_COUNT);
-    }
+// 위치가 domain인 것 안좋음
+// -> infrastructure로
+public interface LottoGenerator {
+    List<Integer> generate();
 }
